requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.12.5
pyyaml==6.0.2
